Add singleton to store reference to Jekyll::Site
The Jekyll::Site object keeps track of static files to add to the
generated _site. It is accessible via Generators, but not Converters.
However, it doesn't make sense to parse through all the posts at the
Generator level when we'll have much more control over that parsing at
the Converter level. So, add a Generator simply to get a reference to
the Jekyll::Site, store it on a singleton, and then access that site in
the Converter to add individual image files.